requirements.txt
annotated-types==0.6.0
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coverage==7.3.2
cssselect==1.2.0
exceptiongroup==1.1.3
fake-useragent==1.3.0
Flask==3.0.0
greenlet==3.0.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.2
parse==1.19.1
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
pytest==7.4.2
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
uWSGI==2.0.23
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.1
zipp==3.17.0
